requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.2
blis==0.7.9
blobfile==2.0.1
bs4==0.0.1
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
confection==0.0.4
contourpy==1.0.7
cx-Oracle==8.3.0
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.6
decorator==5.1.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0-py3-none-any.whl
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.97.0
filelock==3.9.0
Flask==2.2.3
flatbuffers==23.3.3
fonttools==4.38.0
frozenlist==1.3.3
gast==0.4.0
google-auth==2.17.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h11==0.14.0
h5py==3.8.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.12.1
idna==3.4
ipykernel==6.21.2
ipympl==0.9.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyterlab-widgets==3.0.7
keras==2.12.0
kiwisolver==1.4.4
langcodes==3.3.0
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
ml-dtypes==0.0.4
multidict==6.0.4
murmurhash==1.0.9
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.2
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathy==0.10.1
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
plotly==5.13.1
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.17
pydantic==1.10.5
Pygments==2.14.0
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3
# pywin32==305
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
starlette==0.27.0
tenacity==8.2.2
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
# tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
thinc==8.1.7
threadpoolctl==3.1.0
tiktoken==0.3.0
tk==0.1.0
tkcalendar==1.6.1
tokenizers==0.13.2
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.1
typer==0.7.0
typing_extensions==4.6.3
tzdata==2023.3
Unidecode==1.3.6
# urllib3==2.0.3
uvicorn==0.22.0
wasabi==1.1.2
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wrapt==1.14.1
yarl==1.8.2
